Move utilities for blocked vds to new location
