+++ src/nexgen/nxs_write/ed_nxmx_writer.py
@@ -14,7 +14,8 @@ import numpy as np
 from numpy.typing import DTypeLike
 
 from ..nxs_utils import Attenuator, Beam, Detector, Goniometer, Source
-from ..tools.vds_w_tools import image_vds_writer, vds_file_writer
+from ..tools.vds_tools.blocked_mapping import image_vds_writer
+from ..tools.vds_w_tools import vds_file_writer
 from ..utils import coord2mcstas
 from .nxclass_writers import (
     write_NXcoordinate_system_set,
